wip: testing out serverless deployment
