+++ apps/nest-cli/src/commands/command.module.ts
@@ -1,5 +1,8 @@
 import { ImportAdminDataService } from '@codelab/backend/application/admin'
-import { ImportService } from '@codelab/backend/infra/adapter/cli'
+import {
+  ImportService,
+  TerraformService,
+} from '@codelab/backend/infra/adapter/cli'
 import { BullModule } from '@nestjs/bull'
 import type { OnModuleInit } from '@nestjs/common'
 import { Module } from '@nestjs/common'
@@ -12,7 +15,12 @@ import { CommandService } from './command.service'
       name: 'import-admin-data',
     }),
   ],
-  providers: [CommandService, ImportService, ImportAdminDataService],
+  providers: [
+    CommandService,
+    ImportService,
+    ImportAdminDataService,
+    TerraformService,
+  ],
 })
 export class CommandModule implements OnModuleInit {
   constructor(private readonly commandService: CommandService) {}
@@ -20,4 +28,5 @@ export class CommandModule implements OnModuleInit {
   onModuleInit() {
     this.commandService.exec()
   }
+  // https://github.com/codelab-app/platform/blob/2b929d723822dbf9654526deb29dfcef2b7af0e9/libs/backend/infra/adapter/cli/src/commands/import/import.service.ts#L38
 }
