+++ libs/backend/infra/adapter/cli/src/commands/index.ts
@@ -1,10 +1,9 @@
 export * from './codegen/codegen'
 export * from './export/export.command'
-export * from './import/import.command'
 export * from './import/import.service'
 export * from './reset/reset.command'
 export * from './scrape/scrape-antd.command'
 export * from './scrape/scrape-html.command'
 export * from './seed/seed.command'
 export * from './tasks/tasks.command'
-export * from './terraform/terraform.command'
+export * from './terraform/terraform.service'
