+++ terraform/environments/prod/module-redis.tf
@@ -1,0 +1,5 @@
+module "redis" {
+  source = "../../modules/redis"
+
+  digitalocean_token = var.DIGITALOCEAN_TOKEN
+}
