+++ terraform/environments/prod/vars-digitalocean.symlink.tf
@@ -1,0 +1,1 @@
+../shared/vars-digitalocean.tf
