+++ terraform/environments/shared/vars-digitalocean.tf
@@ -1,0 +1,3 @@
+variable "DIGITALOCEAN_TOKEN" {
+  type = string
+}
